FilenameSupport: Move from editor to FX to allow use in action dialog
