+++ org.csstudio.display.builder.editor/src/org/csstudio/display/builder/editor/Messages.java
@@ -27,8 +27,6 @@ public class Messages extends NLS
     public static String AlignTop;
     public static String DistributeHorizontally;
     public static String DistributeVertically;
-    public static String FileTypeAll;
-    public static String FileTypeDisplays;
     public static String Grid;
     public static String LoadDisplay;
     public static String LoadDisplay_TT;
