+++ org.csstudio.display.builder.editor/src/org/csstudio/display/builder/editor/messages.properties
@@ -8,8 +8,6 @@ AlignRight=Right-align selected widgets
 AlignTop=Top-align selected widgets
 DistributeHorizontally=Distribute selected widgets horizontally
 DistributeVertically=Distribute selected widgets vertically
-FileTypeAll=All Files (*.*)
-FileTypeDisplays=Displays (*.bob)
 Grid=Align widgets to Grid
 LoadDisplay=Load display
 LoadDisplay_TT=Load display
