+++ org.csstudio.display.builder.editor/src/org/csstudio/display/builder/editor/properties/PropertyPanelSection.java
@@ -17,7 +17,6 @@ import java.util.Optional;
 import java.util.logging.Level;
 
 import org.csstudio.display.builder.editor.undo.SetMacroizedWidgetPropertyAction;
-import org.csstudio.display.builder.editor.util.FilenameSupport;
 import org.csstudio.display.builder.model.ArrayWidgetProperty;
 import org.csstudio.display.builder.model.DisplayModel;
 import org.csstudio.display.builder.model.MacroizedWidgetProperty;
@@ -40,6 +39,7 @@ import org.csstudio.display.builder.model.properties.RulesWidgetProperty;
 import org.csstudio.display.builder.model.properties.ScriptsWidgetProperty;
 import org.csstudio.display.builder.model.util.ModelResourceUtil;
 import org.csstudio.display.builder.representation.javafx.AutocompleteMenu;
+import org.csstudio.display.builder.representation.javafx.FilenameSupport;
 import org.csstudio.display.builder.util.undo.UndoableActionManager;
 import org.csstudio.javafx.DialogHelper;
 import org.csstudio.javafx.MultiLineInputDialog;
