+++ org.csstudio.display.builder.model/examples/05_actions.bob
@@ -1,13 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <display version="1.0.0">
   <name>Actions</name>
-  <x>-1</x>
-  <y>-1</y>
   <widget type="label" version="2.0.0">
     <name>Label</name>
+    <text>Actions</text>
     <width>281</width>
     <height>31</height>
-    <text>Actions</text>
     <font>
       <font name="Header 1" family="Liberation Sans" style="BOLD" size="22.0">
       </font>
@@ -15,10 +13,10 @@
   </widget>
   <widget type="label" version="2.0.0">
     <name>Label_1</name>
+    <text>Purpose</text>
     <y>41</y>
     <width>331</width>
     <height>30</height>
-    <text>Purpose</text>
     <font>
       <font name="Header 2" family="Liberation Sans" style="BOLD" size="18.0">
       </font>
@@ -26,9 +24,6 @@
   </widget>
   <widget type="label" version="2.0.0">
     <name>Label_4</name>
-    <y>71</y>
-    <width>561</width>
-    <height>180</height>
     <text>Actions are typically defined on the 'Action Button' widget.
 
 Actions can
@@ -38,17 +33,21 @@ Actions can
 - Write a value to a PV
 - Start scripts
 </text>
+    <y>71</y>
+    <width>561</width>
+    <height>180</height>
   </widget>
   <widget type="label" version="2.0.0">
     <name>Label_5</name>
-    <y>261</y>
-    <width>561</width>
-    <height>110</height>
     <text>Actions can be assigned to any widget.
 To invoke them, the user then needs to invoke the context menu
 on the widget (right-click) and select the desired action.
 The use of 'Action Buttons' should be preferred because
 it is more obvious to end users.</text>
+    <x>10</x>
+    <y>370</y>
+    <width>561</width>
+    <height>110</height>
     <font>
       <font name="Comment" family="Liberation Sans" style="ITALIC" size="14.0">
       </font>
@@ -56,11 +55,6 @@ it is more obvious to end users.</text>
   </widget>
   <widget type="action_button" version="2.0.0">
     <name>Action Button</name>
-    <x>381</x>
-    <y>141</y>
-    <width>120</width>
-    <height>30</height>
-    <text>$(actions)</text>
     <actions>
       <action type="open_display">
         <file>01_main.bob</file>
@@ -68,14 +62,12 @@ it is more obvious to end users.</text>
         <description>back to main</description>
       </action>
     </actions>
+    <x>440</x>
+    <y>140</y>
+    <width>120</width>
   </widget>
   <widget type="action_button" version="2.0.0">
     <name>Action Button_1</name>
-    <x>381</x>
-    <y>181</y>
-    <width>120</width>
-    <height>30</height>
-    <text>$(actions)</text>
     <actions>
       <action type="write_pv">
         <pv_name>$(pv_name)</pv_name>
@@ -89,13 +81,30 @@ it is more obvious to end users.</text>
       </action>
     </actions>
     <pv_name>loc://text("some_PV")</pv_name>
+    <x>440</x>
+    <y>251</y>
+    <width>120</width>
   </widget>
   <widget type="textupdate" version="2.0.0">
     <name>Text Update</name>
-    <x>261</x>
-    <y>181</y>
+    <pv_name>loc://text("some_PV")</pv_name>
+    <x>320</x>
+    <y>251</y>
     <width>110</width>
     <height>30</height>
-    <pv_name>loc://text("some_PV")</pv_name>
+  </widget>
+  <widget type="label" version="2.0.0">
+    <name>Label_6</name>
+    <text>Typical actions buttons have exactly one
+action assigned to them.
+Clicking the button invokes that action.
+
+If an action button has multiple actions
+assigned, a drop-down allows invoking
+one of them at a time.</text>
+    <x>10</x>
+    <y>220</y>
+    <width>300</width>
+    <height>140</height>
   </widget>
 </display>
