+++ org.csstudio.display.builder.representation.javafx/src/org/csstudio/display/builder/representation/javafx/messages.properties
@@ -26,6 +26,8 @@ ColorMapDialog_Title=Color Map
 ColorMapDialog_Value=Value (0-255)
 Column=Column
 Copy=Copy
+FileTypeAll=All Files (*.*)
+FileTypeDisplays=Displays (*.bob)
 FontDialog_ExampleText=Example Text
 FontDialog_Family=Custom Font Family:
 FontDialog_Info=Select a predefined font, or a custom font
